+++ web/src/components/form/Modification.vue
@@ -48,6 +48,7 @@ export default {
         {value: 'DELETE', text: 'Delete'},
         {value: 'ARRAY_APPEND', text: 'Append (to array)'},
         {value: 'ARRAY_DELETE', text: 'Delete (from array)'},
+        {value: 'REGEX_PARSE', text: 'Regex capture group'},
       ],
     }
   },